second commit

+++ app/Http/Controllers/PostController.php
@@ -9,7 +9,7 @@ use Illuminate\Support\Facades\Auth;
 class PostController extends Controller
 {
     public function getDashboard(){
-        $posts = Post::all();
+        $posts = Post::orderBy('created_at', 'desc')->get();
         return view('dashboard', ['posts'=>$posts]);
     }
 
@@ -28,7 +28,23 @@ class PostController extends Controller
 
     public function getDeletePost($post_id){
         $post = Post::where('id', $post_id)->first();
+        if(Auth::user()!=$post->user){
+            return redirect()->back();
+        }
         $post->delete();
         return redirect()->route('dashboard')->with(['message'=>'Successfully deleted!']);
+    }
+
+    public function postEditPost(Request $request){
+        $this->validate($request, [
+            'body' => 'required'
+        ]);
+        $post = Post::find($request['postId']);
+        if(Auth::user()!=$post->user){
+            return redirect()->back();
+        }
+        $post->body = $request['body'];
+        $post->update();
+        return response()->json(['new_body'=>$post->body], 200);
     }
 }
